refactor(beads): trim shared link extraction
