+++ src/backends/markdown-grammar.ts
@@ -1,8 +1,7 @@
 import { AxiError } from "../errors.js";
-import type { Dep, Hold, HoldKind, State, Task } from "../model.js";
+import type { Dep, Hold, HoldKind, State, Task, TaskLink } from "../model.js";
 import { HOLD_KINDS } from "../model.js";
-import { deriveLinks } from "../links.js";
-export { deriveLinks } from "../links.js";
+import { isPrUrl } from "../pr-url.js";
 import {
   PUBLIC_FOLLOWUP_KIND,
   assertPublicFollowupTaskState,
@@ -144,6 +143,33 @@ export function leadingKind(title: string): string | undefined {
 
 function titleHasLeadingKind(title: string, kind: string): boolean {
   return leadingKind(title) === kind;
+}
+
+const REPORT_LINK = /\bdata\/\S+?\/report\.md\b/g;
+const GENERIC_URL = /https?:\/\/\S+/g;
+
+function trimUrl(url: string): string {
+  return url.replace(/[).,;]+$/, "");
+}
+
+/** Derive typed links by scanning prose (links live in the prose, not as managed tags). */
+export function deriveLinks(text: string): TaskLink[] {
+  const links: TaskLink[] = [];
+  const seen = new Set<string>();
+  const add = (kind: TaskLink["kind"], raw: string) => {
+    const url = trimUrl(raw);
+    if (seen.has(url)) return;
+    seen.add(url);
+    links.push({ kind, url });
+  };
+  for (const m of text.matchAll(GENERIC_URL)) {
+    if (isPrUrl(trimUrl(m[0]))) add("pr", m[0]);
+  }
+  for (const m of text.matchAll(REPORT_LINK)) add("report", m[0]);
+  for (const m of text.matchAll(GENERIC_URL)) {
+    if (!isPrUrl(trimUrl(m[0]))) add("doc", m[0]);
+  }
+  return links;
 }
 
 export interface ExtractedTags {
