+++ src/commands/crud.ts
@@ -9,8 +9,7 @@ import {
   takeFlag,
 } from "../args.js";
 import { takeBody } from "../body.js";
-import { extractTags } from "../backends/markdown-grammar.js";
-import { deriveLinks } from "../links.js";
+import { deriveLinks, extractTags } from "../backends/markdown-grammar.js";
 import { PR_URL_EXPECTED } from "../pr-url.js";
 import { renderMutation, stateLabel, taskToJson } from "../confirm.js";
 import { requireCtx, type TasksContext } from "../context.js";
